Migrate seen plugins to a component
