+++ flexget/api/__init__.py
@@ -4,5 +4,5 @@ from .app import api_app, api, APIResource, APIClient  # noqa
 from .core import authentication, cached, database, plugins, server, tasks, user, format_checker  # noqa
 from .plugins import (
     series, trakt_lookup, tvdb_lookup, tvmaze_lookup, tmdb_lookup, irc, status, pending_list, pending, entry_list,
-    failed, imdb_lookup, movie_list, rejected, schedule, variables, seen  # noqa
+    failed, imdb_lookup, movie_list, rejected, schedule, variables  # noqa
 )
