+++ flexget/components/seen/__init__.py
@@ -1,4 +1,6 @@
 # Import plugins when component is imported
 from . import api
 from . import cli
-from . import history
+from . import seen
+from . import seen_movies
+from . import seen_info_hash
