+++ flexget/tests/api_tests/test_series_api.py
@@ -1,15 +1,14 @@
 from __future__ import unicode_literals, division, absolute_import
+
 from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import pytest
-from datetime import datetime, timedelta
-
-from flexget.plugins.filter.seen import SeenEntry
 
 from flexget.api.app import base_message
 from flexget.api.plugins.series import ObjectsContainer as OC
 from flexget.api.plugins.tvdb_lookup import ObjectsContainer as tvdb
 from flexget.api.plugins.tvmaze_lookup import ObjectsContainer as tvmaze
+from flexget.components.seen.db import SeenEntry
 from flexget.manager import Session
 from flexget.plugins.filter.series import (
     Series, SeriesTask, Episode, EpisodeRelease, AlternateNames, Season, SeasonRelease
